restarted pop up div
reset most of the popup div style rules, the old ones
fell apart with the new content added, so thought it
best to start from mostly scratch. added a few inline
rules to make it easier to see what i'm doing. stopping
to go fix a bug with the x button