fixed button position for mobile
the old solution of shrinking the popUp div so that the add
to cart button would appear to be outside didn't work with
safari mobile screens. so instead i removed it from the div
in js (and fixed the js function addToCart by removing one
of the parentElements from the target) and then styled the
button to the center by adding display: block. I also
fixed a few reduntant rules in the ul and li of the section