changed switch button method
since I'll likely use gradients for buttons, I didn't want
to waste space by adding a browser checker in the JS to
handle all the various fallbacks. instead, i'll style 2
buttons in the CSS, and hide them one at a time with the
rest of the div and ul. minor changes were made to the JS
and index.html to accomodate this