fixed bug from adding switch buttons
To make the switch buttons work, i jsut redefined the cart and rejects
in the event listeners to be their wrappers, not the uls, however,
I forgot that I used those ul definitions in function parameters.
Meaning, the ul in the function suddenly became a div, which caused
problems. So, I just created new wrapper variables and set the
listeners to those and left the ul variables as they were. easy peazy