reverted to older line design
the blocks were terrible and the gradients looked wrong so I reaaplied
the old line design and now I think it's best. This is a massive commit
unfortunately because I just cut and copied out the old css rules from
a previous commit. I made some minor alterations to update the css rules
to the new code and also kept the gradients for the buttons