add menu centering function
in js i added a function that ensures all the popup menus
stay in the viewport. i edited some of the styles since i
realized the grocerylist popup was identified with a class
and not an id. I also realized calculateTotal function was
missing from the the update button event, so I added that in.