removed testing objects/fixed sale update bug
I removed the list, cart, and rejects starting list items (and object indexes)
as well as fixed a bug in the updateItemPrice function where the calculated
sale as a decimal point was put into the user input box. This was confusing to
see, and if it wasn't fixed with each press of the button, the sale would decrease
by 100% each time. I made the slace calculated value its own variable to fix it