*fixed* rejects, changed total calc, moved total over
Not sure what the issue with rejects was, items wouldn't delete, but then
after adding and removing an alert, it suddenly worked fine again.
Must've done something I didn't realize. Once that was *fixed* I updated
and gave rejects a mini finalPrice function inside it to display the
final price, not the price at the index of the item, when it moves
back into cart. Then i realized that Cart Total was using the wrong
price for each item, since it was just adding up all the numbers in
groceryObj.prices. Now, it calculates a final price using the sale and
amounts, and then adds all those final prices together. finally, since
I used the left side of the screen to position the total (so it won't
look weird if opened on a browser) i realized as the total increases, the
tab moves over to the right and off screen. I just scooted it over to the left