addName now adds to localStorage
Addname function now adds item name to the groceryListObj, which
in turn is placed in localStorage.groceryList. In the IIFE I have
placed the first part of what will be a localStorage check function.
right now it only checks groceryList. I'll add more as I go so any
breakdowns are caught immediately