fixed pastToGrocery issue
realized that adding the span tag around the text changed
the element layer, so fixed the js to compensate. Also changed
the event listener to look for the span's class, not the li's
class (also realized i had originally set the spans to have id's,
not classes. So i set them to have classes to correct the mistake)