fixed array emptying issue
learned that when you set an array to array = [], it actually
makes a duplicate, local array that is empty, while leaving the
global array's value untouched. so I switched all instances of
this to array.length = 0, except for the initial checkStorage
(becuase those empty arrays were forcing the value to update
by returning an array of the various empty arrays, and then
the global variables were set off that unnamed array's empty
lists. This solution also works to solve the duplicate array
issue